% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap}
\alias{dap}
\alias{dapc}
\alias{dapr}
\alias{dapc_if}
\alias{dapr_if}
\title{Data frame apply function}
\usage{
dap(.data, .f, ...)

dapc(.data, .f, ...)

dapr(.data, .f, ...)

dapc_if(.data, .predicate, .f, ...)

dapr_if(.data, .predicate, .f, ...)
}
\arguments{
\item{.data}{Data frame input.}

\item{.f}{Function to apply to element (columns or rows)}

\item{...}{Other args passed to function call}

\item{.predicate}{Logical vector of expression evaluating to a logical vector}

\item{.predicate}{Logical vector of expression evaluating to a logical vector}
}
\value{
A data frame
}
\description{
Applies functions to columns (default) of data frame
}
