% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilap.R
\name{ilap}
\alias{ilap}
\title{Iterator list apply}
\usage{
ilap(.data, .f, ...)
}
\arguments{
\item{.data}{The sequence length, used as ".i" in the following expression}

\item{.f}{Function or formula call that assumes integer element is .i}

\item{...}{Other values passed to function call.}
}
\value{
A list
}
\description{
Iterate over sequence length of input and return list(s)
}
\examples{

## return string list
ilap(1:10, ~ paste0(letters[.i], rev(LETTERS)[.i]))

## return list of columns
ilap(mtcars, ~ c(row.names(mtcars)[.i], mtcars$wt[.i]))

}
