% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lap.R, R/vap.R
\name{lap}
\alias{lap}
\alias{vap_chr}
\alias{vap_dbl}
\alias{vap_lgl}
\title{List apply}
\usage{
lap(.x, .f, ...)

vap_chr(.x, .f, ...)

vap_dbl(.x, .f, ...)

vap_lgl(.x, .f, ...)
}
\arguments{
\item{.x}{Input vector}

\item{.f}{Function or formula call that assumes element is .x}

\item{.x}{Input vector}

\item{.f}{Function or formula call that assumes element is .x}
}
\value{
A list

A character vector

A double vector

A logical vector
}
\description{
Iterate over input and return list(s)

Iterate over input and return character(s)

Iterate over input and return double(s)

Iterate over input and return logical(s)
}
\examples{

## return string list
lap(letters, ~ paste0(.x, "."))

## return list of columns
lap(mtcars[1:5, ], as.character)


## character
vap_chr(letters, ~ paste0(.x, "."))

## double
vap_dbl(rnorm(4), round, 3)

## logical
vap_lgl(letters, ~ .x \%in\% c("a", "e", "i", "o", "u"))

}
