% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapc.R
\name{vapc}
\alias{vapc}
\alias{vapc_chr}
\alias{vapc_dbl}
\alias{vapc_lgl}
\alias{vapc_int}
\title{vapr: Vector apply-to-column functions}
\usage{
vapc_chr(.data, .f, ...)

vapc_dbl(.data, .f, ...)

vapc_lgl(.data, .f, ...)

vapc_int(.data, .f, ...)
}
\arguments{
\item{.data}{Input object–must be two-dimensional (e.g., matrix, data.frame)
–over which the second dimension (columns) will be iterated}

\item{.f}{Action to apply to each element of \code{.data}. The action can be
  articulated in one of the four following ways:
\enumerate{
\item supplying a function object (e.g., \code{mean})
\item defining a function (in-line; e.g., \code{function(x) mean(x)})
\item specifying a formula-like call where '.x' is assumed to be the iterated
  over element of \code{.data} (e.g., \code{~ mean(.x)})
\item providing a name or position of \code{.data} to return (e.g.,
  \code{1}, \code{"varname"}, etc.)
}}

\item{...}{Other values passed to function call.}
}
\value{
A character vector

A numeric vector

A logical vector

An integer vector
}
\description{
Functions that apply expressions to the columns of input objects and return
atomic vectors e.g., numeric (double), integer, character, logical.

vapc_chr: Iterate over input \strong{columns} and return \strong{character(s)}

vapc_dbl: Iterate over input \strong{columns} and return \strong{numeric(s)}

vapc_lgl: Iterate over input \strong{columns} and return \strong{logical(s)}

vapc_int: Iterate over input \strong{columns} and return \strong{integer(s)}
}
\seealso{
\code{\link{lap}}  \code{\link{dap}}

Other vap: \code{\link{vapr}}, \code{\link{vap}}
}
\concept{vap}
