% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapr.R
\name{vapr}
\alias{vapr}
\alias{vapr_chr}
\alias{vapr_dbl}
\alias{vapr_lgl}
\alias{vapr_int}
\title{vapr: Vector apply-to-row functions}
\usage{
vapr_chr(.data, .f, ...)

vapr_dbl(.data, .f, ...)

vapr_lgl(.data, .f, ...)

vapr_int(.data, .f, ...)
}
\arguments{
\item{.data}{Input object–must be two-dimensional (e.g., matrix, data.frame)
–over which the first dimension (rows) will be iterated}

\item{.f}{Action to apply to each element of \code{.data}. The action can be
  articulated in one of the four following ways:
\enumerate{
\item supplying a function object (e.g., \code{mean})
\item defining a function (in-line; e.g., \code{function(x) mean(x)})
\item specifying a formula-like call where '.x' is assumed to be the iterated
  over element of \code{.data} (e.g., \code{~ mean(.x)})
\item providing a name or position of \code{.data} to return (e.g.,
  \code{1}, \code{"varname"}, etc.)
}}

\item{...}{Other values passed to function call.}
}
\value{
A character vector

A numeric vector

A logical vector

An integer vector
}
\description{
Functions that apply expressions to the rows of input objects and return
atomic vectors e.g., numeric (double), integer, character, logical.

vapr_chr: Iterate over input \strong{rows} and return \strong{character(s)}

vapr_dbl: Iterate over input \strong{rows} and return \strong{numeric(s)}

vapr_lgl: Iterate over input \strong{rows} and return \strong{logical(s)}

vapr_int: Iterate over input \strong{rows} and return \strong{integer(s)}
}
\seealso{
\code{\link{lap}}  \code{\link{dap}}

Other vap: \code{\link{vapc}}, \code{\link{vap}}
}
\concept{vap}
