% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilap.R
\name{ilap}
\alias{ilap}
\title{Iterator list apply}
\usage{
ilap(.data, .f, ...)
}
\arguments{
\item{.data}{Object representing the sequence length, used as ".i" in the .f
expression. If matrix or data frame, sequence length will be determined by
the number of columns. If integer, then it will be passed directly onto
function call, otherwise the object will be converted into a sequence from
1 to the length of the object.}

\item{.f}{Function to apply to each element (as an integer position) of input
object. This can be written as a single function name e.g., \code{mean}, a
formula-like function call where '.i' is assumed to be the integer position
of input data object e.g., \code{~ mean(mtcars[[.i]])}, or an in-line
function definition e.g., \code{function(x) mean(mtcars[[x]])}.}

\item{...}{Other values passed to function call.}
}
\value{
A list
}
\description{
ilap: Iterate over sequence length of input and return list(s)
}
\examples{

## return string list
ilap(1:10, ~ paste0(letters[.i], rev(LETTERS)[.i]))

## return list of columns
ilap(mtcars, ~ c(row.names(mtcars)[.i], mtcars$wt[.i]))

}
\seealso{
Other lap: \code{\link{lap}}
}
\concept{lap}
