% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lap.R
\name{lap}
\alias{lap}
\alias{dapc}
\alias{dapr}
\alias{dapc_if}
\title{List apply}
\usage{
lap(.data, .f, ...)

dapc(.data, .f, ...)

dapr(.data, .f, ...)

dapc_if(.data, .predicate, .f, ...)
}
\arguments{
\item{.data}{Input vector}

\item{.f}{Function or formula call that assumes element is .x}

\item{...}{Other args passed to function call}

\item{.predicate}{Logical vector or expression evaluated to expression.}
}
\value{
A list
}
\description{
Iterate over input and return list(s)
}
\examples{

## return string list
lap(letters, ~ paste0(.x, "."))

## return list of columns
lap(mtcars[1:5, ], as.character)

}
