% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vap.R
\name{vap_chr}
\alias{vap_chr}
\alias{vap_dbl}
\alias{vap_lgl}
\alias{vap_int}
\title{Vector apply character}
\usage{
vap_chr(.data, .f, ...)

vap_dbl(.data, .f, ...)

vap_lgl(.data, .f, ...)

vap_int(.data, .f, ...)
}
\arguments{
\item{.data}{Input vector}

\item{.f}{Function or formula call that assumes element is .data}

\item{...}{Args}
}
\value{
A character vector

A double vector

A logical vector

A integer vector
}
\description{
Iterate over input and return character(s)

Iterate over input and return double(s)

Iterate over input and return logical(s)

Iterate over input and return integer(s)
}
\examples{

## character
vap_chr(letters, ~ paste0(.x, "."))

## double
vap_dbl(rnorm(4), round, 3)

## logical
vap_lgl(letters, ~ .x \%in\% c("a", "e", "i", "o", "u"))

## integer
vap_int(rnorm(5), ~ as.integer(.x))

}
