% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vap.R
\name{vap_chr}
\alias{vap_chr}
\alias{vap_dbl}
\alias{vap_lgl}
\title{Vector apply character}
\usage{
vap_chr(.x, .f, ...)

vap_dbl(.x, .f, ...)

vap_lgl(.x, .f, ...)
}
\arguments{
\item{.x}{Input vector}

\item{.f}{Function or formula call that assumes element is .x}

\item{...}{Args}
}
\value{
A character vector

A double vector

A logical vector
}
\description{
Iterate over input and return character(s)

Iterate over input and return double(s)

Iterate over input and return logical(s)
}
\examples{

## character
vap_chr(letters, ~ paste0(.x, "."))

## double
vap_dbl(rnorm(4), round, 3)

## logical
vap_lgl(letters, ~ .x \%in\% c("a", "e", "i", "o", "u"))

}
